# apps/alpha-web/frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "JobDetail": {
        "additionalProperties": false,
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "lines": {
            "items": {
              "type": "string"
            },
            "title": "Lines",
            "type": "array"
          },
          "n_lines": {
            "title": "N Lines",
            "type": "integer"
          },
          "returncode": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returncode"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "command",
          "kind",
          "status",
          "created_at",
          "run_id",
          "returncode",
          "n_lines",
          "lines"
        ],
        "title": "JobDetail",
        "type": "object"
      },
      "JobStatus": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "JobStatus",
        "type": "object"
      },
      "JobSummary": {
        "additionalProperties": false,
        "properties": {
          "command": {
            "title": "Command",
            "type": "string"
          },
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "n_lines": {
            "title": "N Lines",
            "type": "integer"
          },
          "returncode": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Returncode"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "command",
          "kind",
          "status",
          "created_at",
          "run_id",
          "returncode",
          "n_lines"
        ],
        "title": "JobSummary",
        "type": "object"
      },
      "LaunchRequest": {
        "description": "Launch body: a command path (e.g. ``\"backtest run\"``) + its remaining args, or a bare\n``args`` string (empty ``command``) for the free-form console.",
        "properties": {
          "args": {
            "default": "",
            "title": "Args",
            "type": "string"
          },
          "command": {
            "default": "",
            "title": "Command",
            "type": "string"
          }
        },
        "title": "LaunchRequest",
        "type": "object"
      },
...
      "OptionCurve": {
        "additionalProperties": false,
        "properties": {
          "days": {
            "title": "Days",
            "type": "number"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/OptionCurvePoint"
            },
            "minItems": 2,
            "title": "Points",
            "type": "array"
          },
          "rate": {
            "title": "Rate",
            "type": "number"
          },
          "strike": {
            "title": "Strike",
            "type": "number"
          },
          "vol": {
            "title": "Vol",
            "type": "number"
          }
        },
        "required": [
          "strike",
          "vol",
          "days",
          "rate",
          "kind",
          "points"
        ],
        "title": "OptionCurve",
        "type": "object"
      },
      "OptionCurvePoint": {
        "additionalProperties": false,
        "properties": {
          "delta": {
            "title": "Delta",
            "type": "number"
          },
          "gamma": {
            "title": "Gamma",
            "type": "number"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "spot": {
            "title": "Spot",
            "type": "number"
          },
          "theta": {
            "title": "Theta",
            "type": "number"
          },
          "vega": {
            "title": "Vega",
            "type": "number"
          }
        },
        "required": [
          "spot",
          "price",
          "delta",
          "gamma",
          "vega",
          "theta"
        ],
        "title": "OptionCurvePoint",
        "type": "object"
      },
      "OptionGreeks": {
        "additionalProperties": false,
        "properties": {
          "days": {
            "title": "Days",
            "type": "number"
          },
          "delta": {
            "title": "Delta",
            "type": "number"
          },
          "gamma": {
            "title": "Gamma",
            "type": "number"
          },
          "implied_vol": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implied Vol"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "market_price": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Market Price"
          },
          "price": {
            "title": "Price",
            "type": "number"
          },
          "rate": {
            "title": "Rate",
            "type": "number"
          },
          "rho": {
            "title": "Rho",
            "type": "number"
          },
          "spot": {
            "title": "Spot",
            "type": "number"
          },
          "strike": {
            "title": "Strike",
            "type": "number"
          },
          "theta": {
            "title": "Theta",
            "type": "number"
          },
          "vega": {
            "title": "Vega",
            "type": "number"
          },
          "vol": {
            "title": "Vol",
            "type": "number"
          }
        },
        "required": [
          "spot",
          "strike",
          "rate",
          "vol",
          "days",
          "kind",
          "price",
          "delta",
          "gamma",
          "vega",
          "theta",
          "rho"
        ],
        "title": "OptionGreeks",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Project ALPHA \u2014 Workstation",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/jobs": {
      "get": {
        "description": "All known jobs (live + this-session-finished), newest first.",
        "operationId": "list_jobs_api_jobs_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobSummary"
                  },
                  "title": "Response List Jobs Api Jobs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "description": "Launch ``alpha <command> <args>`` as a background job. Returns its id + initial status.",
        "operationId": "launch_job_api_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LaunchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Launch Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}": {
      "delete": {
        "description": "Cancel a running job (idempotent no-op if already finished).",
        "operationId": "cancel_job_api_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "description": "A single job's status + its buffered output lines (for a late-opening panel).",
        "operationId": "get_job_api_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/jobs/{job_id}/stream": {
      "get": {
        "description": "SSE of a job's output; a reconnect with ``Last-Event-ID`` replays only the missed lines.",
        "operationId": "stream_job_api_jobs__job_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Job",
        "tags": [
          "jobs"
        ]
      }
    },
    "/api/options/curve": {
      "get": {
        "description": "Price + greeks across a spot range for the greeks-vs-spot chart.",
        "operationId": "options_curve_api_options_curve_get",
        "parameters": [
          {
            "in": "query",
            "name": "strike",
            "required": true,
            "schema": {
              "title": "Strike",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "vol",
            "required": true,
            "schema": {
              "title": "Vol",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30.0,
              "title": "Days",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "rate",
            "required": false,
            "schema": {
              "default": 0.05,
              "title": "Rate",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "call",
              "title": "Kind",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "width",
            "required": false,
            "schema": {
              "default": 0.5,
              "title": "Width",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "points",
            "required": false,
            "schema": {
              "default": 41,
              "maximum": 501,
              "minimum": 2,
              "title": "Points",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionCurve"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Options Curve",
        "tags": [
          "options"
        ]
      }
    },
    "/api/options/greeks": {
      "get": {
        "description": "Price + delta/gamma/vega/theta/rho for one European option.",
        "operationId": "options_greeks_api_options_greeks_get",
        "parameters": [
          {
            "in": "query",
            "name": "spot",
            "required": true,
            "schema": {
              "title": "Spot",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "strike",
            "required": true,
            "schema": {
              "title": "Strike",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "vol",
            "required": true,
            "schema": {
              "title": "Vol",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30.0,
              "title": "Days",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "rate",
            "required": false,
            "schema": {
              "default": 0.05,
              "title": "Rate",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "call",
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionGreeks"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Options Greeks",
        "tags": [
          "options"
        ]
      }
    },
    "/api/options/iv": {
      "get": {
        "description": "The implied volatility that reprices the option to ``price`` (+ greeks at that vol).",
        "operationId": "options_iv_api_options_iv_get",
        "parameters": [
          {
            "in": "query",
            "name": "spot",
            "required": true,
            "schema": {
              "title": "Spot",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "strike",
            "required": true,
            "schema": {
              "title": "Strike",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "price",
            "required": true,
            "schema": {
              "title": "Price",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 30.0,
              "title": "Days",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "rate",
            "required": false,
            "schema": {
              "default": 0.05,
              "title": "Rate",
              "type": "number"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "default": "call",
              "title": "Kind",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OptionGreeks"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Options Iv",
        "tags": [
          "options"
        ]
      }
    },
...
  }
}